fixed some errors; moved components;
