+++ app/components/avatar.js
