+++ app/components/protectRoutes.js
